Increased ticks data type.

+++ Assets/TimeUtils.cs
@@ -2,10 +2,10 @@ namespace SS {
   public static class TimeUtils {
     public const ushort CIT_CYCLE = 280;
 
-    public static int SecondsToSlowTicks(double seconds) => (int)(60.0 * seconds);
-    public static int SecondsToFastTicks(double seconds) => (int)(CIT_CYCLE * seconds);
+    public static long SecondsToSlowTicks(double seconds) => (long)(60.0 * seconds);
+    public static long SecondsToFastTicks(double seconds) => (long)(CIT_CYCLE * seconds);
 
-    public static ushort SecondsToTimestamp(double seconds) => (ushort)(((uint)(CIT_CYCLE * seconds) >> 4) & 0xFFFF);
+    public static ushort SecondsToTimestamp(double seconds) => (ushort)(((ulong)(CIT_CYCLE * seconds) >> 4) & 0xFFFF);
     public static ushort FastTicksToTimestamp(uint ticks) => (ushort)((ticks >> 4) & 0xFFFF);
     public static int TimestampToFastTicks(ushort timestamp) => timestamp << 4;
   }
